[ios] Use `UIImage` instead of `CGDataProvider` to read PNG and read orientation
This solves the issue where certain images were being rotated. We can access `imageOrientation` from `UIImage` when we read in file, which we couldn't do with the `CGImage`. If the image orientation is `.right` (portrait photos), we set `imageNeedsRotation` and include the transformation.

There's definitely a cleaner way to do this, but I couldn't find one after spending a few hours so this'll do for now.